requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Security
cryptography==41.0.8
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Docker
docker==6.1.3

# Cloud providers
boto3==1.34.0
google-cloud-compute==1.14.1
google-cloud-sql==0.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Visualization
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0

# WebSocket
flask-socketio==5.3.6
python-socketio==5.10.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0